css/css-sizing/aspect-ratio/replaced-element-036.html

<!DOCTYPE html>
<title>CSS aspect-ratio: img</title>
<link rel="author" title="Jake Archibald" href="jakearchibald@google.com" />
<link rel="help" href="https://drafts.csswg.org/css-sizing-4/#aspect-ratio" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<img
  id="test"
  src="error.png"
  style="width: 100px; aspect-ratio: 1/2; background: green; display: block"
  alt="Should be 200px high"
/>

<script>
  onload = () => {
    test(() => {
      assert_equals(document.getElementById("test").offsetHeight, 200);
    });
  };
</script>
